+++ helper.py
@@ -5,7 +5,7 @@ from rdflib import BNode, URIRef
 from bs4 import BeautifulSoup
 
 from config import Config
-from triplestore import Triplestore
+# from triplestore import Triplestore
 
 import re
 from urllib.parse import quote_plus
@@ -30,22 +30,22 @@ def render_concept_tree(html_doc):
     return soup
 
 
-def get_triplestore_created_time():
-    """Get the string message of the last time the local graph cache was created."""
-
-    MSG = 'Last updated {}.'
-    for created_time in Config.g.objects(Triplestore.THIS_GRAPH, DCTERMS.created):
-        created_time = created_time.toPython()
-        now = datetime.now()
-        now -= timedelta(minutes=1)
-
-        last_updated = (datetime.now() - created_time).seconds // 60
-        if not last_updated:
-            LAST_UPDATE_VALUE = 'just now'
-        else:
-            LAST_UPDATE_VALUE = f'{last_updated} minutes ago'
-
-        return MSG.format(LAST_UPDATE_VALUE)
+# def get_triplestore_created_time():
+#     """Get the string message of the last time the local graph cache was created."""
+#
+#     MSG = 'Last updated {}.'
+#     for created_time in Config.g.objects(Triplestore.THIS_GRAPH, DCTERMS.created):
+#         created_time = created_time.toPython()
+#         now = datetime.now()
+#         now -= timedelta(minutes=1)
+#
+#         last_updated = (datetime.now() - created_time).seconds // 60
+#         if not last_updated:
+#             LAST_UPDATE_VALUE = 'just now'
+#         else:
+#             LAST_UPDATE_VALUE = f'{last_updated} minutes ago'
+#
+#         return MSG.format(LAST_UPDATE_VALUE)
 
 
 def uri_label(uri):